node/requirements.txt
psutil==6.0.0
requests==2.32.3
pyjwt==2.8.0
cryptography==43.0.0
